requirements.txt
duckdb>=1.1,<2.0
psycopg[binary]>=3.2,<4.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
python-dateutil>=2.9,<3.0
pytz>=2024.1
